// AngularSPA/src/app/shared/models/MovieDetailsModel.ts
// Generated by https://quicktype.io

import { GenreModel } from "./GenreModel";

export interface MovieDetailsModel {
    id:               number;
    title:            string;
    overview:         string;
    tagline:          string;
    budget:           number;
    revenue:          number;
    imdbUrl:          string;
    tmdbUrl:          string;
    posterUrl:        string;
    backdropUrl:      string;
    originalLanguage: string;
    releaseDate:      string;
    runTime:          number;
    price:            number;
    rating:           number;
    genres:           GenreModel[];
    trailers:         Trailer[];
    casts:            Cast[];
}

export interface Cast {
    id:          number;
    name:        string;
    profilePath: string;
    character:   string;
}

export interface Trailer {
    trailerUrl: string;
    name:       string;
}