// AngularSPA/src/app/shared/models/ActiveUserModel.ts
// Generated by https://quicktype.io

export interface ActiveUserModel {
    nameid:      string;
    email:       string;
    family_name: string;
    given_name:  string;
    language:    string;
    isAdmin:     string;
    nbf:         number;
    exp:         number;
    iat:         number;
    iss:         string;
    aud:         string;
}